requirements.txt
tensorflow
matplotlib
numpy
opencv_python
opencv_python_headless
Pillow
plantcv